Don't compile director-test cleanup tests on Windows
The test fixtures hard-coded RFC3339 timestamps like
"director-test-2025-01-10T10:00:00Z.txt", which fail on Windows CI
because `:` is reserved in filenames. Besides, real test files are only ever created on Linux caches.
